move Hall_parameter to plasmapy.formulary.dimensionless
